==> Part 1 - Insta csv production.ipynb <==
# Instabang .csv Production Using Pandas

import os
import pandas as pd
import numpy as np
import csv

from sklearn.model_selection import train_test_split
from pathlib import Path

pwd

with open('./moonmountainman/moomountainman_text/moonmountainman_caption_index', 'w', encoding='utf-8') as out_file:
    csv_out = csv.writer(out_file)
    csv_out.writerow(['FileName', 'Content'])
    for fileName in Path('./moonmountainman/moomountainman_text/').glob('*.txt'):
        lines = [ ]
        with open(str(fileName.absolute()),'rb') as one_text:
            for line in one_text.readlines():
                lines.append(line.decode(encoding='utf-8').strip())
        csv_out.writerow([str(fileName),' '.join(lines)])

df_text = pd.read_csv('./moonmountainman/moomountainman_text/moonmountainman_caption_index')

di = []
for row in df_text.FileName:
    d = row.split('/')
#     d = row.split('.txt')
    di.append(d[2])

df_text = df_text.assign(FileName=di)

df_text.sort_values(by='FileName', axis=0, inplace=True)
df_text.head()

df_image = pd.read_excel('./moonmountainman/moonmountainman_image/moonmountainman_image_index.xlsx', header=0)
df_image.head()

df_image['Captions'] = [df_text.Content[df_text['FileName'] == x].values for x in df_image['Image FileName'].map(lambda x: x.split('UTC')[0]+'UTC.txt')]
df_image.head()

df_image['Captions']= df_image['Captions'].apply(lambda x: x.tolist()[0] if len(x)>0 else '')
df_image.head()

train, test = train_test_split(df_image, test_size=0.1)

train.drop(columns='Captions', inplace=True, axis=1)

test.drop(columns='Captions', inplace=True, axis=1)

df_image.to_csv('moonmountainman_index.csv', header=False, index=False)

train.to_csv('moonmountainman_train.csv', header=False, index=False, sep=':')

test.to_csv('moonmountainman_test.csv', header=False, index=False, sep=':')

# Joining All .csv's

Loading all index .csv's

df_beautifulxtravel = pd.read_csv('./beautifulxtravel/beautifultravel_text/beautifulxtravel_index.csv', header=None)

df_chloebarryhang = pd.read_csv('./chloebarryhang/chloebarryhang_text/chloebarryhang_index.csv', header=None)

df_explorerssaurus = pd.read_csv('./explorerssaurus_/explorerssaurus_text/explorerssaurus_index.csv', header=None)

df_jesswandering = pd.read_csv('./jesswandering/jesswandering_text/jesswandering_index.csv', header=None)

df_lostleblanc = pd.read_csv('./lostleblanc/lostleblanc_text/lostleblanc_index.csv', header=None)

df_moonmountainman = pd.read_csv('./moonmountainman_index.csv', header=None)

df_theblondeabroad = pd.read_csv('./theblondeabroad/theblondabroad_text/theblondeabroad_index.csv', header=None)

df_watchluke = pd.read_csv('./watchluke/watchluke_text/watchluke_index.csv', header=None)

df_mikesugianto = pd.read_csv('./mikesugianto/mikesugianto_text/mikesugianto_index.csv', header=None)

df_jukkarisikko = pd.read_csv('./jukkarisikko/jukkarisikko_text/jukkarisikko_index.csv', header=None)

instabang_index = df_beautifulxtravel.append(df_chloebarryhang, ignore_index=True)

instabang_index = instabang_index.append(df_explorerssaurus, ignore_index=True)

instabang_index = instabang_index.append(df_jesswandering, ignore_index=True)

instabang_index = instabang_index.append(df_lostleblanc, ignore_index=True)

instabang_index = instabang_index.append(df_moonmountainman, ignore_index=True)

instabang_index = instabang_index.append(df_theblondeabroad, ignore_index=True)

instabang_index = instabang_index.append(df_watchluke, ignore_index=True)

instabang_index = instabang_index.append(df_mikesugianto, ignore_index=True)

instabang_index = instabang_index.append(df_jukkarisikko, ignore_index=True)

instabang_index.to_csv('instabang_index.txt', header=False, index=False, sep='\\', mode='w')

instabang_index_df = pd.read_csv('instabang_index.txt', sep='\\', header=None)

df_beautifulxtravel_train = pd.read_csv('./beautifulxtravel/beautifultravel_text/beautifulxtravel_train.csv', header=None)

df_chloebarryhang_train = pd.read_csv('./chloebarryhang/chloebarryhang_text/chloebarryhang_train.csv', header=None)

df_explorerssaurus_train = pd.read_csv('./explorerssaurus_/explorerssaurus_text/explorerssaurus_train.csv', header=None)

df_jesswandering_train = pd.read_csv('./jesswandering/jesswandering_text/jesswandering_train.csv', header=None)

df_lostleblanc_train = pd.read_csv('./lostleblanc/lostleblanc_text/lostleblanc_train.csv', header=None)

df_moonmountainman_train = pd.read_csv('./moonmountainman/moomountainman_text/moonmountainman_train.csv', header=None)

df_theblondeabroad_train = pd.read_csv('./theblondeabroad/theblondabroad_text/theblondeabroad_train.csv', header=None)

df_watchluke_train = pd.read_csv('./watchluke/watchluke_text/watchluke_train.csv', header=None)

df_mikesugianto_train = pd.read_csv('./mikesugianto/mikesugianto_text/mikesugianto_train.csv', header=None)

df_jukkarisikko_train = pd.read_csv('./jukkarisikko/jukkarisikko_text/jukkarisikko_train.csv', header=None)

df_beautifulxtravel_test = pd.read_csv('./beautifulxtravel/beautifultravel_text/beautifulxtravel_test.csv', header=None)

df_chloebarryhang_test = pd.read_csv('./chloebarryhang/chloebarryhang_text/chloebarryhang_test.csv', header=None)

df_explorerssaurus_test = pd.read_csv('./explorerssaurus_/explorerssaurus_text/explorerssaurus_test.csv', header=None)

df_jesswandering_test = pd.read_csv('./jesswandering/jesswandering_text/jesswandering_test.csv', header=None)

df_lostleblanc_test = pd.read_csv('./lostleblanc/lostleblanc_text/lostleblanc_test.csv', header=None)

df_moonmountainman_test = pd.read_csv('./moonmountainman/moomountainman_text/moonmountainman_test.csv', header=None)

df_theblondeabroad_test = pd.read_csv('./theblondeabroad/theblondabroad_text/theblondeabroad_test.csv', header=None)

df_watchluke_test = pd.read_csv('./watchluke/watchluke_text/watchluke_test.csv', header=None)

df_mikesugianto_test = pd.read_csv('./mikesugianto/mikesugianto_text/mikesugianto_test.csv', header=None)

df_jukkarisikko_test = pd.read_csv('./jukkarisikko/jukkarisikko_text/jukkarisikko_test.csv', header=None)

instabang_train = df_beautifulxtravel_train.append(df_chloebarryhang_train, ignore_index=True)

instabang_train = instabang_train.append(df_explorerssaurus_train, ignore_index=True)

instabang_train = instabang_train.append(df_jesswandering_train, ignore_index=True)

instabang_train = instabang_train.append(df_lostleblanc_train, ignore_index=True)

instabang_train = instabang_train.append(df_moonmountainman_train, ignore_index=True)

instabang_train = instabang_train.append(df_theblondeabroad_train, ignore_index=True)

instabang_train = instabang_train.append(df_watchluke_train, ignore_index=True)

instabang_train = instabang_train.append(df_mikesugianto_train, ignore_index=True)

instabang_train = instabang_train.append(df_jukkarisikko_train, ignore_index=True)

instabang_train.drop(instabang_train.columns[1], axis=1, inplace=True)

instabang_test = df_beautifulxtravel_test.append(df_chloebarryhang_test, ignore_index=True)

instabang_test = instabang_test.append(df_explorerssaurus_test, ignore_index=True)

instabang_test = instabang_test.append(df_jesswandering_test, ignore_index=True)

instabang_test = instabang_test.append(df_lostleblanc_test, ignore_index=True)

instabang_test = instabang_test.append(df_moonmountainman_test, ignore_index=True)

instabang_test = instabang_test.append(df_theblondeabroad_test, ignore_index=True)

instabang_test = instabang_test.append(df_watchluke_test, ignore_index=True)

instabang_test = instabang_test.append(df_mikesugianto_test, ignore_index=True)

instabang_test = instabang_test.append(df_jukkarisikko_test, ignore_index=True)

instabang_train.to_csv('instabang_train.txt', header=False, index=False, mode='w')

instabang_test.to_csv('instabang_test.txt', header=False, index=False, mode='w')
